[#286] report generation improvements
