+++ src/main/java/de/bonndan/nivio/output/docs/OwnersReportGenerator.java
@@ -1,8 +1,10 @@
 package de.bonndan.nivio.output.docs;
 
-import de.bonndan.nivio.model.*;
-import de.bonndan.nivio.output.Color;
-import de.bonndan.nivio.output.FormatUtils;
+import de.bonndan.nivio.assessment.Assessment;
+import de.bonndan.nivio.model.Component;
+import de.bonndan.nivio.model.GroupedBy;
+import de.bonndan.nivio.model.Item;
+import de.bonndan.nivio.model.Landscape;
 import de.bonndan.nivio.output.LocalServer;
 import de.bonndan.nivio.output.icons.IconService;
 import j2html.tags.ContainerTag;
@@ -11,8 +13,7 @@ import org.springframework.lang.Nullable;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.stream.Collectors;
 
 import static j2html.TagCreator.*;
 
@@ -23,50 +24,33 @@ public class OwnersReportGenerator extends HtmlGenerator {
     }
 
     @Override
-    public String toDocument(@NonNull final Landscape landscape, @Nullable final SearchConfig searchConfig) {
-        return writeLandscape(landscape, searchConfig != null ? searchConfig.getSearchTerm() : null);
-    }
+    public String toDocument(@NonNull final Landscape landscape, @NonNull final Assessment assessment, @Nullable final SearchConfig searchConfig) {
 
-    private String writeLandscape(Landscape landscape, String searchTerm) {
+        String searchTerm = searchConfig != null ? searchConfig.getSearchTerm() : null;
 
-        List<Item> search = searchTerm != null ? new ArrayList<>(landscape.search(searchTerm)) : new ArrayList<>(landscape.getItems().all());
+        List<Item> search = new ArrayList<>(searchTerm != null ? landscape.search(searchTerm) : landscape.getItems().all());
         return html(
                 getHead(landscape),
                 body(
                         h1("Owner Report: " + landscape.getName()),
                         br(),
-                        rawHtml(writeOwnerGroups(GroupedBy.by(Component::getOwner, search)))
+                        rawHtml(writeOwnerGroups(GroupedBy.by(Component::getOwner, search), assessment))
                 )
         ).renderFormatted();
     }
 
-    private String writeOwnerGroups(GroupedBy ownerGroups) {
+    private String writeOwnerGroups(GroupedBy ownerGroups, Assessment assessment) {
         final StringBuilder builder = new StringBuilder();
-        ownerGroups.getAll().forEach((owner, landscapeItems) -> {
+        ownerGroups.getAll().forEach((owner, items) -> {
             builder.append(
-                    h2(rawHtml(owner)).attr("class", "rounded").render()
+                    h2("Owner: " + owner).attr("class", "rounded").render()
             );
-            builder.append(writeGroups(GroupedBy.by(Item::getGroup, landscapeItems)).render());
+            items.forEach(item -> {
+                builder.append(writeItem(item, assessment).render());
+            });
         });
 
         return builder.toString();
-    }
-
-    private ContainerTag writeGroups(GroupedBy groups) {
-        List<ContainerTag> collect = new ArrayList<>();
-        groups.getAll().entrySet().forEach(entry -> collect.add(writeGroup(entry)));
-        return ul().with(collect);
-    }
-
-    private ContainerTag writeGroup(Map.Entry<String, List<Item>> services) {
-        return li().with(services.getValue().stream().map(this::writeItem));
-    }
-
-    private ContainerTag writeItem(Item item) {
-        String groupColor = "#" + Color.nameToRGB(item.getGroup(), Color.GRAY);
-
-        return div(rawHtml("<span style=\"color: " + groupColor + "\">&#9899;</span> " + FormatUtils.nice(item.getGroup()) + ": " + item.toString() + " (" + item.getFullyQualifiedIdentifier().toString() + ")"));
-
     }
 
 }
